Dependency of phrases on `LANG_SERVER`
Fixed building of message parts based on the client language.
Previously, this led to the construction and sending of the same type of message with language differences.

Now the main language for generating messages is the server language.

I don't see this as a problem at the moment, but provided a workaround.